requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via httpx
certifi==2025.6.15
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.2
    # via requests
cryptography==45.0.4
    # via
    #   igwn-auth-utils
    #   ligo-gracedb
    #   scitokens
deprecation==2.1.0
    # via
    #   postgrest
    #   storage3
gotrue==2.12.2
    # via supabase
h11==0.16.0
    # via httpcore
h2==4.2.0
    # via httpx
hpack==4.1.0
    # via h2
httpcore==1.0.9
    # via httpx
httpx[http2]==0.28.1
    # via
    #   gotrue
    #   postgrest
    #   storage3
    #   supabase
    #   supafunc
hyperframe==6.1.0
    # via h2
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
igwn-auth-utils==1.4.0
    # via ligo-gracedb
iniconfig==2.1.0
    # via pytest
ligo-gracedb==2.14.2
    # via -r requirements.in
numpy==2.3.1
    # via pandas
packaging==25.0
    # via
    #   deprecation
    #   pytest
pandas==2.3.0
    # via -r requirements.in
pluggy==1.6.0
    # via pytest
postgrest==1.1.1
    # via supabase
pycparser==2.22
    # via cffi
pydantic==2.11.7
    # via
    #   gotrue
    #   postgrest
pydantic-core==2.33.2
    # via pydantic
pygments==2.19.2
    # via pytest
pyjwt==2.10.1
    # via
    #   gotrue
    #   scitokens
pytest==8.4.1
    # via pytest-mock
pytest-mock==3.14.1
    # via gotrue
python-dateutil==2.9.0.post0
    # via
    #   pandas
    #   storage3
pytz==2025.2
    # via pandas
realtime==2.5.3
    # via supabase
requests==2.32.4
    # via
    #   igwn-auth-utils
    #   ligo-gracedb
    #   scitokens
safe-netrc==1.0.1
    # via igwn-auth-utils
scitokens==1.8.1
    # via igwn-auth-utils
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
storage3==0.12.0
    # via supabase
strenum==0.4.15
    # via supafunc
supabase==2.16.0
    # via -r requirements.in
supafunc==0.10.1
    # via supabase
typing-extensions==4.14.0
    # via
    #   pydantic
    #   pydantic-core
    #   realtime
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via requests
websockets==15.0.1
    # via realtime
